requirements.txt
MAPIE==0.9.1
matplotlib==3.9.2
ml_collections==0.1.1
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
protobuf==4.25.3
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
skfolio==0.4.1
SQLAlchemy==2.0.35
statsmodels==0.14.2
streamlit==1.38.0
streamlit_shadcn_ui==0.1.18
torch==2.3.0
torchdiffeq==0.2.4
tqdm==4.66.5
yfinance==0.2.44
